requirements.txt
streamlit>=1.37
python-dotenv>=1.0
sentence-transformers>=2.7
pinecone-client>=3.2
cohere>=5.5
groq>=0.9
numpy>=1.26
pytest>=8.2
